refactor(sharemodal): delete barrel file and update imports
