--- public/app/features/dashboard/components/ShareModal/index.ts
@@ -1,2 +1,0 @@
-export { ShareModal, addDashboardShareTab, addPanelShareTab } from './ShareModal';
-export * from './types';
