+++ public/app/features/dashboard/utils/panel.ts
@@ -6,7 +6,7 @@ import appEvents from 'app/core/app_events';
 import config from 'app/core/config';
 import { LS_PANEL_COPY_KEY, PANEL_BORDER } from 'app/core/constants';
 import store from 'app/core/store';
-import { ShareModal } from 'app/features/dashboard/components/ShareModal';
+import { ShareModal } from 'app/features/dashboard/components/ShareModal/ShareModal';
 import { DashboardModel } from 'app/features/dashboard/state/DashboardModel';
 import { PanelModel } from 'app/features/dashboard/state/PanelModel';
 import { UnlinkModal } from 'app/features/dashboard-scene/scene/UnlinkModal';
